Implemented Delete & Update and tested the functionalities for Product Section

+++ src/main/java/ro/zah/auction/dto/ProductDTO.java
@@ -15,5 +15,5 @@ public class ProductDTO {
     private Double basePrice;
     private Double settledPrice;
     private List<Bid> bids;
-    private int auctionId;
+    private long auctionId;
 }
